requirements.txt
watchdog
Kivy~=2.2.0.dev0
kivymd~=1.2.0.dev0
Pillow~=9.4.0
requests~=2.28.2
multitasking~=0.0.11
javaobj-py3==0.4.3
pycryptodomex==3.17
protobuf==4.21.12